tests/smart-context/requirements.txt
sentence-transformers>=2.2.0
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
json5>=0.9.0
pytest>=7.0.0
pytest-html>=3.0.0
psycopg2-binary>=2.9.0

